alteracoes aula 6, tratamento de erros e exceptions

+++ src/main/java/com/residencia/academia/controller/TurmaController.java
@@ -15,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.residencia.academia.entity.Turma;
+import com.residencia.academia.exception.NoSuchElementFoundException;
 import com.residencia.academia.service.TurmaService;
 
 @RestController
@@ -31,34 +32,75 @@ public class TurmaController {
 
 	@GetMapping("/{id}")
 	public ResponseEntity<Turma> findTurmaById(@PathVariable Integer id) {
-		return new ResponseEntity<>(turmaService.findTurmaById(id), HttpStatus.OK);
-
+//		return new ResponseEntity<>(turmaService.findTurmaById(id), HttpStatus.OK);
+		
+//		Turma turma = turmaService.findTurmaById(id);
+//		if(turmaService.findTurmaById(id) == null) {
+//			return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+//			return ResponseEntity.notFound().build();
+//		} else {
+//			return new ResponseEntity<>(turmaService.findTurmaById(id), HttpStatus.OK);
+//			return ResponseEntity.ok(turmaService.findTurmaById(id));
+//		}
+		
+		Turma turma = turmaService.findTurmaById(id);
+		if(null == turma) {
+			throw new NoSuchElementFoundException("Não foi possível encontrar a turma de id = " + id);
+		} else {
+			return new ResponseEntity<>(turmaService.findTurmaById(id), HttpStatus.OK);
+		}
+		
 	}
 
 	@PostMapping
 	public ResponseEntity<Turma> saveTurma(@RequestBody Turma turma) {
-//		
-//		SimpleDateFormat sdf = new SimpleDateFormat("HH:mm");
-//		String horario = sdf.format(turma.getHorarioTurma());
-//		horario.
-//		
 		return new ResponseEntity<>(turmaService.saveTurma(turma), HttpStatus.CREATED);
 
 	}
 
 	@PutMapping
 	public ResponseEntity<Turma> updateTurma(@RequestBody Turma turma) {
-		return new ResponseEntity<>(turmaService.updateTurma(turma), HttpStatus.OK);
+//		return new ResponseEntity<>(turmaService.updateTurma(turma), HttpStatus.OK);
+//		return ResponseEntity.ok(turmaService.updateTurma(turma));
 
+		if(turmaService.updateTurma(turma) == null) {
+			return ResponseEntity.notFound().build();
+		} else {
+			return ResponseEntity.ok().body(turmaService.updateTurma(turma));
+		}
+		
 	}
 
+//	@DeleteMapping("/{id}")
+//	public ResponseEntity<String> deleteTurmaById(@PathVariable Integer id) {
+//		turmaService.deleteTurmaById(id);
+//		return new ResponseEntity<>("Turma deletada com sucesso.", HttpStatus.NO_CONTENT);
+//		return new ResponseEntity<>("Turma deletada com sucesso.", HttpStatus.OK);
+//		
+//	}
+	
 	@DeleteMapping("/{id}")
 	public ResponseEntity<String> deleteTurma(@PathVariable Integer id) {
-		turmaService.deleteTurma(id);
-		return new ResponseEntity<>("Turma deletada com sucesso.", HttpStatus.OK);
-//		return new ResponseEntity<>("Turma deletada com sucesso.", HttpStatus.NO_CONTENT);
+		Turma turma = turmaService.findTurmaById(id);
+		if(null == turma)
+			throw new NoSuchElementFoundException("Não foi possível excluir a Turma, "
+					+ "pois não foi "
+					+ "encontrada uma turma com o id " + id);
 
+		turmaService.deleteTurma(id);
+		return new ResponseEntity<>("", HttpStatus.OK);
 	}
+	
+	
+//	@DeleteMapping("/delete-com-conferencia/{id}")
+//	public ResponseEntity<String> deleteTurmaComConferencia(@PathVariable Integer id) {
+//		Boolean verificacao = turmaService.deleteTurmaComConferencia(id);
+//		if(verificacao) {
+//			return new ResponseEntity<>("", HttpStatus.OK);
+//		} else {
+//			throw new NoSuchElementFoundException("Não foi possível excluir a turma. Não foi encontrada a turma de id = " + id);
+//		}
+//	}
 
 //	@DeleteMapping
 //	public void deleteTurma(@RequestBody Turma turma) {
